[build] threads/activities ui updated

+++ lib/ui/routes/activity/activity.dart
@@ -5,23 +5,28 @@ import 'package:myaniapp/graphql.dart';
 import 'package:myaniapp/graphql/__generated/graphql/schema.graphql.dart';
 import 'package:myaniapp/graphql/__generated/ui/routes/activity/activity.graphql.dart';
 import 'package:myaniapp/graphql/__generated/ui/routes/home/activities/activities.graphql.dart';
+import 'package:myaniapp/providers/user.dart';
 import 'package:myaniapp/ui/common/activity_card.dart';
 import 'package:myaniapp/ui/common/banner_ad.dart';
 import 'package:myaniapp/ui/common/comment.dart';
 import 'package:myaniapp/ui/common/dialogs/delete.dart';
 import 'package:myaniapp/ui/common/graphql_error.dart';
+import 'package:myaniapp/ui/common/hidden_floating_action_button.dart';
 import 'package:myaniapp/ui/common/markdown/markdown.dart';
 import 'package:myaniapp/ui/common/markdown_editor.dart';
 import 'package:myaniapp/ui/common/pagination.dart';
 import 'package:myaniapp/utils/require_login.dart';
 
 class ActivityPage extends ConsumerWidget {
-  const ActivityPage({super.key, required this.id});
+  ActivityPage({super.key, required this.id});
 
   final int id;
+  final ScrollController _controller = ScrollController();
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
+    var user = ref.watch(userProvider);
+
     return Stack(
       alignment: Alignment.bottomCenter,
       children: [
@@ -67,52 +72,46 @@ class ActivityPage extends ConsumerWidget {
                 pageInfo: result.parsedData!.replies!.pageInfo!,
                 child: Scaffold(
                   appBar: AppBar(),
-                  floatingActionButton: FloatingActionButton(
-                    onPressed: requireLogin(
-                      ref,
-                      'reply',
-                      () => showMarkdownEditor(
-                        context,
-                        onSave: (text) {
-                          if (text.isNotEmpty) {
-                            client.value.mutate$SaveActivityReply(
-                              Options$Mutation$SaveActivityReply(
-                                variables: Variables$Mutation$SaveActivityReply(
-                                  activityId: id,
-                                  text: text,
+                  floatingActionButton: HiddenFloatingActionButton(
+                    scrollController: _controller,
+                    child: FloatingActionButton(
+                      onPressed: requireLogin(
+                        ref,
+                        'reply',
+                        () => showMarkdownEditor(
+                          context,
+                          onSave: (text) {
+                            if (text.isNotEmpty) {
+                              client.value.mutate$SaveActivityReply(
+                                Options$Mutation$SaveActivityReply(
+                                  variables:
+                                      Variables$Mutation$SaveActivityReply(
+                                    activityId: id,
+                                    text: text,
+                                  ),
+                                  onCompleted: (p0, p1) => refetch!(),
                                 ),
-                                onCompleted: (p0, p1) => refetch!(),
-                              ),
-                            );
-                          }
-                        },
+                              );
+                            }
+                          },
+                        ),
                       ),
+                      child: const Icon(Icons.reply),
                     ),
-                    child: const Icon(Icons.reply),
                   ),
                   body: RefreshIndicator.adaptive(
                     onRefresh: refetch!,
                     child: CustomScrollView(
+                      controller: _controller,
                       slivers: [
                         SliverToBoxAdapter(
-                          child: Column(
-                            children: [
-                              ActivityCard(
-                                activity: result.parsedData!.activity!,
-                                onDelete: () => context.pop(true),
-                                inActivity: true,
-                              ),
-                              const SizedBox(
-                                height: 10,
-                              ),
-                              Text(
-                                'Replies ${result.parsedData?.replies?.activityReplies?.length ?? 0}',
-                                style: Theme.of(context)
-                                    .textTheme
-                                    .titleLarge
-                                    ?.copyWith(fontWeight: FontWeight.bold),
-                              )
-                            ],
+                          child: Padding(
+                            padding: const EdgeInsets.only(bottom: 20),
+                            child: ActivityCard(
+                              activity: result.parsedData!.activity!,
+                              onDelete: () => context.pop(true),
+                              inActivity: true,
+                            ),
                           ),
                         ),
                         SliverList.builder(
@@ -127,7 +126,7 @@ class ActivityPage extends ConsumerWidget {
                               avatarUrl: reply.user?.avatar?.large,
                               createdAt: reply.createdAt,
                               username: reply.user?.name,
-                              leading: Row(
+                              footer: Row(
                                 children: [
                                   IconButton(
                                     icon: Row(
@@ -157,41 +156,53 @@ class ActivityPage extends ConsumerWidget {
                                       ),
                                     ),
                                   ),
-                                  ActivityPopupMenu(
-                                    text: reply.text,
-                                    userId: reply.userId,
-                                    onDelete: () =>
-                                        showDeleteDialog(context).then(
-                                      (value) {
-                                        if (value == true) {
-                                          client.value
-                                              .mutate$DeleteActivityReply(
-                                            Options$Mutation$DeleteActivityReply(
-                                              variables:
-                                                  Variables$Mutation$DeleteActivityReply(
-                                                id: reply.id,
+                                  if (user.value?.id == reply.userId)
+                                    IconButton(
+                                      onPressed: () => showMarkdownEditor(
+                                        context,
+                                        text: reply.text,
+                                        onSave: (text) {
+                                          if (text.length > 2) {
+                                            client.value
+                                                .mutate$SaveActivityReply(
+                                              Options$Mutation$SaveActivityReply(
+                                                variables:
+                                                    Variables$Mutation$SaveActivityReply(
+                                                  id: reply.id,
+                                                  text: text,
+                                                ),
                                               ),
-                                              onCompleted: (p0, p1) =>
-                                                  refetch(),
-                                            ),
-                                          );
-                                        }
-                                      },
+                                            );
+                                          }
+                                        },
+                                      ),
+                                      icon: const Icon(Icons.edit),
                                     ),
-                                    onEdit: (text) {
-                                      if (text.length > 2) {
-                                        client.value.mutate$SaveActivityReply(
-                                          Options$Mutation$SaveActivityReply(
-                                            variables:
-                                                Variables$Mutation$SaveActivityReply(
-                                              id: reply.id,
-                                              text: text,
-                                            ),
-                                          ),
-                                        );
-                                      }
-                                    },
-                                  ),
+                                  if (user.value?.id == reply.userId)
+                                    IconButton(
+                                      onPressed: () =>
+                                          showDeleteDialog(context).then(
+                                        (value) {
+                                          if (value == true) {
+                                            client.value
+                                                .mutate$DeleteActivityReply(
+                                              Options$Mutation$DeleteActivityReply(
+                                                variables:
+                                                    Variables$Mutation$DeleteActivityReply(
+                                                  id: reply.id,
+                                                ),
+                                                onCompleted: (p0, p1) =>
+                                                    refetch(),
+                                              ),
+                                            );
+                                          }
+                                        },
+                                      ),
+                                      icon: const Icon(
+                                        Icons.delete,
+                                        color: Colors.red,
+                                      ),
+                                    ),
                                 ],
                               ),
                             );
